Move router types to routers.rs
